refactor: updated app logo
