+++ src/components/icons/Logo.tsx
@@ -1,39 +1,27 @@
 import { SVGProps } from 'react';
 
-interface Props extends SVGProps<SVGSVGElement> {}
-
-function Logo(props: Props) {
+function Logo(props: SVGProps<SVGSVGElement>) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="151"
-      height="40"
+      width={284}
+      height={46}
       fill="none"
-      aria-hidden="true"
-      aria-label="Peer logo"
       {...props}
     >
       <path
-        d="M46.5508 34.396C46.5508 37.1574 44.3122 39.396 41.5508 39.396H25.3446V35.3589C25.3446 33.7021 26.6878 32.3589 28.3446 32.3589H46.5508V34.396Z"
-        fill="#61F3E1"
+        fill="#000"
+        d="M4.013 38V7.455h10.321c2.387 0 4.345.407 5.877 1.223 1.53.805 2.664 1.914 3.4 3.326.736 1.412 1.104 3.017 1.104 4.817 0 1.8-.368 3.396-1.104 4.788-.736 1.392-1.864 2.485-3.386 3.28-1.52.786-3.465 1.18-5.831 1.18H6.042v-3.342h8.232c1.631 0 2.944-.238 3.938-.716 1.004-.477 1.73-1.153 2.178-2.028.457-.885.686-1.939.686-3.162 0-1.223-.23-2.292-.686-3.207-.458-.914-1.189-1.62-2.193-2.118-1.004-.507-2.332-.76-3.982-.76H7.712V38H4.013Zm14.378-13.722L25.908 38h-4.295l-7.398-13.722h4.176ZM40.531 38h-3.878L47.87 7.455h3.818L62.903 38h-3.877l-9.128-25.713h-.239L40.531 38Zm1.432-11.932h15.63v3.281h-15.63v-3.28ZM79.76 7.455V38h-3.699V7.455h3.699ZM95.576 38V7.455h3.7v27.264h14.198V38H95.576Zm48.73-22.91c-.179-1.51-.905-2.684-2.178-3.519-1.272-.835-2.833-1.253-4.683-1.253-1.352 0-2.535.219-3.55.656-1.004.438-1.789 1.04-2.356 1.805a4.324 4.324 0 0 0-.835 2.61c0 .816.194 1.517.581 2.103a5.16 5.16 0 0 0 1.522 1.447c.616.378 1.262.691 1.939.94.676.238 1.297.432 1.864.581l3.102.835c.796.21 1.681.498 2.655.865.984.368 1.924.87 2.819 1.507a7.672 7.672 0 0 1 2.237 2.416c.587.985.88 2.193.88 3.624 0 1.651-.432 3.142-1.297 4.475-.856 1.332-2.108 2.391-3.759 3.177-1.641.785-3.634 1.178-5.981 1.178-2.187 0-4.082-.353-5.682-1.059-1.591-.706-2.844-1.69-3.759-2.953-.905-1.263-1.417-2.73-1.536-4.4h3.818c.1 1.153.487 2.108 1.163 2.864.687.745 1.552 1.302 2.596 1.67 1.054.358 2.187.537 3.4.537 1.412 0 2.68-.229 3.803-.686 1.124-.467 2.014-1.114 2.67-1.939.656-.835.985-1.81.985-2.923 0-1.014-.284-1.84-.851-2.476-.566-.637-1.312-1.154-2.237-1.551a23.524 23.524 0 0 0-2.998-1.044l-3.758-1.074c-2.387-.686-4.276-1.666-5.668-2.938-1.392-1.273-2.088-2.939-2.088-4.997 0-1.71.462-3.201 1.387-4.474.935-1.283 2.188-2.277 3.759-2.983 1.581-.716 3.346-1.074 5.294-1.074 1.969 0 3.719.353 5.25 1.059 1.532.696 2.745 1.65 3.64 2.864.905 1.213 1.382 2.59 1.431 4.13h-3.579ZM163.179 37V22.454h1.676v5.37h.143c.123-.19.293-.431.511-.725.222-.298.54-.563.952-.795.416-.237.98-.355 1.69-.355.918 0 1.728.23 2.429.689.701.46 1.248 1.11 1.641 1.953s.589 1.837.589 2.983c0 1.155-.196 2.157-.589 3.004-.393.843-.938 1.496-1.634 1.96-.696.46-1.499.69-2.408.69-.7 0-1.262-.117-1.683-.349-.421-.236-.746-.504-.973-.802a10.777 10.777 0 0 1-.525-.753h-.199V37h-1.62Zm1.648-5.454c0 .823.121 1.55.362 2.18.242.625.594 1.115 1.059 1.47.464.35 1.032.526 1.704.526.701 0 1.286-.185 1.754-.554.474-.374.829-.876 1.066-1.506.241-.634.362-1.34.362-2.116 0-.768-.118-1.459-.355-2.074-.232-.62-.585-1.11-1.058-1.47-.469-.365-1.059-.547-1.769-.547-.682 0-1.255.172-1.719.518-.464.34-.814.82-1.051 1.435-.237.61-.355 1.323-.355 2.137Zm15.236 9.545c-.284 0-.538-.024-.76-.071a1.993 1.993 0 0 1-.462-.128l.426-1.477c.408.104.767.142 1.08.113.312-.028.589-.168.831-.419.246-.246.471-.646.675-1.2l.312-.852-4.034-10.966h1.818l3.012 8.693h.113l3.012-8.693h1.818l-4.631 12.5c-.208.563-.466 1.03-.774 1.4a2.898 2.898 0 0 1-1.073.83c-.402.18-.857.27-1.363.27Z"
       />
       <path
-        d="M0 11.0703H41.5501C44.3115 11.0703 46.5501 13.3089 46.5501 16.0703V18.1074H0V11.0703Z"
-        fill="#61F3E1"
+        fill="#000"
+        fillRule="evenodd"
+        d="M268.767 22.476h-8.897v1.09c0 .627.507 1.134 1.133 1.134h7.571v1.207h-6.861v2.288h6.861v1.271h-7.571c-.625 0-1.133.508-1.133 1.134v1.09h8.897a2.225 2.225 0 0 0 2.225-2.216v-4.836h-.001a2.225 2.225 0 0 0-2.224-2.162Zm-13.462.108h-8.897v1.09c0 .626.507 1.134 1.133 1.134h7.57v1.207h-6.861v2.288h6.861v1.271h-7.57c-.626 0-1.134.508-1.134 1.134v1.09h8.898a2.224 2.224 0 0 0 2.224-2.224v-4.828h-.001a2.223 2.223 0 0 0-2.223-2.162Zm20.89-.108h6.179c.898 0 1.626.728 1.626 1.626v2.601c0 .898-.728 1.626-1.626 1.626h-1.645c.114.094.223.196.323.308l2.948 3.27h-1.728a2.897 2.897 0 0 1-2.153-.959l-2.362-2.619h-1.562v3.577h-2.277v-9.43h2.277Zm5.528 3.577h-5.528v-1.3h5.528v1.3Zm-39.914-3.577H233.529v9.235h2.229v-3.503h6.051c.88 0 1.592-.713 1.592-1.592v-2.548c0-.88-.712-1.592-1.592-1.592Zm-6.051 3.503h5.414v-1.274h-5.414v1.274Z"
+        clipRule="evenodd"
       />
       <path
-        d="M0 5.60401C0 2.84258 2.23858 0.604004 5 0.604004H21.2062V4.64105C21.2062 6.29791 19.863 7.64106 18.2062 7.64106H0V5.60401Z"
         fill="#61F3E1"
-      />
-      <path
-        d="M0 21.5366H46.5501V28.5737H5.00001C2.23858 28.5737 0 26.3351 0 23.5737V21.5366Z"
-        fill="#61F3E1"
-      />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M70.2473 11.0601H58.7768H54.5508V12.2675V17.7009V21.9269V28.5677H58.7768V21.9269H70.2473C71.9144 21.9269 73.2658 20.5753 73.2658 18.9083V14.0786C73.2658 12.4116 71.9144 11.0601 70.2473 11.0601ZM58.7768 17.7009H69.0399V15.286H58.7768V17.7009ZM122.736 11.0601H104.486V14.1436C104.486 14.7695 104.993 15.2768 105.619 15.2768H120.985V17.5638H107.979V21.9011H120.985V24.3127H105.619C104.993 24.3127 104.486 24.8201 104.486 25.446V28.5295H122.736C124.301 28.5295 125.569 27.2611 125.569 25.6963V24.3127H125.569V15.2768H125.57V13.8933C125.57 12.3285 124.301 11.0601 122.736 11.0601ZM97.2124 11.2645H78.9617V14.348C78.9617 14.9739 79.4691 15.4813 80.095 15.4813H95.4625V17.7683H82.4555V22.1055H95.4625V24.5196H80.095C79.4691 24.5196 78.9617 25.027 78.9617 25.6529V28.7364H97.2124C98.7771 28.7364 100.046 27.4679 100.046 25.9032V24.5319H100.046V15.365H100.046V14.0977C100.046 12.533 98.7771 11.2645 97.2124 11.2645ZM135.434 11.0601H147.147C148.85 11.0601 150.23 12.4402 150.23 14.1425V19.0744C150.23 20.7767 148.85 22.1569 147.147 22.1569H144.113L150.229 28.94H145.798C144.977 28.94 144.195 28.5919 143.645 27.9821L138.393 22.1569H135.434V28.9382H131.119V22.1569V17.8414V12.293V11.0601H135.434ZM145.914 17.8414H135.434V15.3755H145.914V17.8414Z"
-        fill="black"
+        d="M226.458 33.685a3.697 3.697 0 0 1-3.697 3.697h-7.445v-.697a3 3 0 0 1 3-3h8.142ZM202 22.5h20.761a3.698 3.698 0 0 1 3.698 3.697H202V22.5ZM202 20.697A3.697 3.697 0 0 1 205.697 17h7.445v.697a3 3 0 0 1-3 3H202ZM202 28h24.459v3.697h-20.762A3.697 3.697 0 0 1 202 28Z"
       />
     </svg>
   );
