+++ src/components/icons/LogoOld.tsx
@@ -1,0 +1,42 @@
+import { SVGProps } from 'react';
+
+interface Props extends SVGProps<SVGSVGElement> {}
+
+function Logo(props: Props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="151"
+      height="40"
+      fill="none"
+      aria-hidden="true"
+      aria-label="Peer logo"
+      {...props}
+    >
+      <path
+        d="M46.5508 34.396C46.5508 37.1574 44.3122 39.396 41.5508 39.396H25.3446V35.3589C25.3446 33.7021 26.6878 32.3589 28.3446 32.3589H46.5508V34.396Z"
+        fill="#61F3E1"
+      />
+      <path
+        d="M0 11.0703H41.5501C44.3115 11.0703 46.5501 13.3089 46.5501 16.0703V18.1074H0V11.0703Z"
+        fill="#61F3E1"
+      />
+      <path
+        d="M0 5.60401C0 2.84258 2.23858 0.604004 5 0.604004H21.2062V4.64105C21.2062 6.29791 19.863 7.64106 18.2062 7.64106H0V5.60401Z"
+        fill="#61F3E1"
+      />
+      <path
+        d="M0 21.5366H46.5501V28.5737H5.00001C2.23858 28.5737 0 26.3351 0 23.5737V21.5366Z"
+        fill="#61F3E1"
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M70.2473 11.0601H58.7768H54.5508V12.2675V17.7009V21.9269V28.5677H58.7768V21.9269H70.2473C71.9144 21.9269 73.2658 20.5753 73.2658 18.9083V14.0786C73.2658 12.4116 71.9144 11.0601 70.2473 11.0601ZM58.7768 17.7009H69.0399V15.286H58.7768V17.7009ZM122.736 11.0601H104.486V14.1436C104.486 14.7695 104.993 15.2768 105.619 15.2768H120.985V17.5638H107.979V21.9011H120.985V24.3127H105.619C104.993 24.3127 104.486 24.8201 104.486 25.446V28.5295H122.736C124.301 28.5295 125.569 27.2611 125.569 25.6963V24.3127H125.569V15.2768H125.57V13.8933C125.57 12.3285 124.301 11.0601 122.736 11.0601ZM97.2124 11.2645H78.9617V14.348C78.9617 14.9739 79.4691 15.4813 80.095 15.4813H95.4625V17.7683H82.4555V22.1055H95.4625V24.5196H80.095C79.4691 24.5196 78.9617 25.027 78.9617 25.6529V28.7364H97.2124C98.7771 28.7364 100.046 27.4679 100.046 25.9032V24.5319H100.046V15.365H100.046V14.0977C100.046 12.533 98.7771 11.2645 97.2124 11.2645ZM135.434 11.0601H147.147C148.85 11.0601 150.23 12.4402 150.23 14.1425V19.0744C150.23 20.7767 148.85 22.1569 147.147 22.1569H144.113L150.229 28.94H145.798C144.977 28.94 144.195 28.5919 143.645 27.9821L138.393 22.1569H135.434V28.9382H131.119V22.1569V17.8414V12.293V11.0601H135.434ZM145.914 17.8414H135.434V15.3755H145.914V17.8414Z"
+        fill="black"
+      />
+    </svg>
+  );
+}
+
+export default Logo;
