+++ src/components/icons/index.ts
@@ -1,6 +1,7 @@
 export { default as BankIcon } from './BankIcon';
 export { default as ConfirmationModalIcon } from './ConfirmationModalIcon';
 export { default as Logo } from './Logo';
+export { default as LogoOld } from './LogoOld';
 export { default as LogoIcon } from './LogoIcon';
 export { default as MatchingIcon } from './MatchingIcon';
 export { default as MatchedIcon } from './MatchedIcon';
